+++ instrumentation/github.com/Shopify/sarama/otelsarama/example/producer/producer.go
@@ -38,7 +38,10 @@ var (
 )
 
 func main() {
-	tp := example.InitTracer()
+	tp, err := example.InitTracer()
+	if err != nil {
+		log.Fatal(err)
+	}
 	defer func() {
 		if err := tp.Shutdown(context.Background()); err != nil {
 			log.Printf("Error shutting down tracer provider: %v", err)
@@ -54,7 +57,10 @@ func main() {
 	brokerList := strings.Split(*brokers, ",")
 	log.Printf("Kafka brokers: %s", strings.Join(brokerList, ", "))
 
-	producer := newAccessLogProducer(brokerList)
+	producer, err := newAccessLogProducer(brokerList)
+	if err != nil {
+		log.Fatal(err)
+	}
 
 	rand.Seed(time.Now().Unix())
 
@@ -75,14 +81,14 @@ func main() {
 	successMsg := <-producer.Successes()
 	log.Println("Successful to write message, offset:", successMsg.Offset)
 
-	err := producer.Close()
+	err = producer.Close()
 	if err != nil {
 		span.SetStatus(codes.Error, err.Error())
 		log.Fatalln("Failed to close producer:", err)
 	}
 }
 
-func newAccessLogProducer(brokerList []string) sarama.AsyncProducer {
+func newAccessLogProducer(brokerList []string) (sarama.AsyncProducer, error) {
 	config := sarama.NewConfig()
 	config.Version = sarama.V2_5_0_0
 	// So we can know the partition and offset of messages.
@@ -90,7 +96,7 @@ func newAccessLogProducer(brokerList []string) sarama.AsyncProducer {
 
 	producer, err := sarama.NewAsyncProducer(brokerList, config)
 	if err != nil {
-		log.Fatalln("Failed to start Sarama producer:", err)
+		return nil, fmt.Errorf("starting Sarama producer: %w", err)
 	}
 
 	// Wrap instrumentation
@@ -103,5 +109,5 @@ func newAccessLogProducer(brokerList []string) sarama.AsyncProducer {
 		}
 	}()
 
-	return producer
+	return producer, nil
 }